move UI ts interfaces from UI components, fix webpack optimization, add some ts-types for webpack modules separate babel-loader from ts-loader add correct lint mode in package.json add separate babel-loader & ts-loader fix eslint warnings & errors
