+++ frontend/babel.config.json
@@ -1,6 +1,7 @@
 {
   "presets": [
     "@babel/preset-env",
-    ["@babel/preset-react", {"runtime": "automatic"}]
+    ["@babel/preset-react", {"runtime": "automatic"}],
+    "@babel/preset-typescript"
   ]
 }
