+++ frontend/components/App.tsx
@@ -2,6 +2,7 @@ import BaseForm from './forms/BaseForm'
 import PostsList from './PostsList'
 
 const App = () => {
+    const id = 2
 	return (
 		<>
 			<PostsList/>
