+++ frontend/package.json
@@ -4,11 +4,12 @@
   "main": "src/index.tsx",
   "type": "module",
   "scripts": {
-    "build": "cross-env NODE_ENV=production webpack --mode production",
+    "build": "cross-env NODE_OPTIONS=\"--loader ts-node/esm\" NODE_ENV=production webpack --mode production",
     "start": "cross-env NODE_OPTIONS=\"--loader ts-node/esm\" NODE_ENV=development webpack serve --mode development --open",
-    "test": "ts-node --esm",
-    "eslint": "cross-env NODE_ENV=eslint_errors webpack serve --mode development  --open",
-    "stats": "cross-env NODE_ENV=bundle_analyzer webpack --mode production --json > stats.json && webpack-bundle-analyzer stats.json",
+    "eslint": "cross-env NODE_OPTIONS=\"--loader ts-node/esm\" NODE_ENV=eslint_errors webpack serve  --mode development  --open",
+    "lint": "cross-env NODE_OPTIONS=\"--loader ts-node/esm\" NODE_ENV=eslint_errors eslint . --ext=ts,tsx",
+    "start-babel": "cross-env NODE_OPTIONS=\"--loader ts-node/esm\" NODE_ENV=with_babel webpack serve --mode development --open",
+    "stats": "cross-env NODE_OPTIONS=\"--loader ts-node/esm\" NODE_ENV=bundle_analyzer webpack --mode production --json > stats.json && webpack-bundle-analyzer stats.json",
     "clean": "rd /s /q dist"
   },
   "keywords": [
@@ -31,7 +32,7 @@
   },
   "devDependencies": {
     "@babel/core": "^7.22.19",
-    "@babel/eslint-parser": "^7.22.15",
+    "@babel/eslint-parser": "^7.24.7",
     "@babel/preset-env": "^7.22.15",
     "@babel/preset-react": "^7.22.15",
     "@babel/preset-typescript": "^7.24.7",
@@ -42,8 +43,8 @@
     "@types/webpack": "^5.28.2",
     "@types/webpack-bundle-analyzer": "^4.7.0",
     "@types/webpack-dev-server": "^4.7.2",
-    "@typescript-eslint/eslint-plugin": "^6.10.0",
-    "@typescript-eslint/parser": "^6.10.0",
+    "@typescript-eslint/eslint-plugin": "^7.15.0",
+    "@typescript-eslint/parser": "^7.15.0",
     "babel-loader": "^9.1.3",
     "clean-webpack-plugin": "^4.0.0",
     "copy-webpack-plugin": "^11.0.0",
@@ -52,9 +53,9 @@
     "css-minimizer-webpack-plugin": "^5.0.1",
     "csv-loader": "^3.0.5",
     "dotenv-webpack": "^8.0.1",
-    "eslint": "^8.49.0",
-    "eslint-plugin-react": "^7.33.2",
-    "eslint-webpack-plugin": "^4.0.1",
+    "eslint": "^8.56.0",
+    "eslint-plugin-react": "^7.34.3",
+    "eslint-webpack-plugin": "^4.2.0",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.5.3",
     "less": "^4.2.0",
