+++ frontend/src/index.tsx
@@ -5,4 +5,5 @@ import 'bootstrap/dist/js/bootstrap.js';
 import App from '../components/App'
 
 const root = ReactDOM.createRoot(document.getElementById('app'));
+
 root.render(<App />);
