+++ frontend/tsconfig.json
@@ -5,7 +5,7 @@
     "noEmit": false,
     "noEmitOnError": false,
     "module": "ESNext",
-    "target": "es5",
+    "target": "es2015",
     "jsx": "react-jsx",
     "allowJs": true,
     "moduleResolution": "node",
