+++ frontend/webpack.config.ts
@@ -1,7 +1,8 @@
 import path from 'path'
 import webpack from "webpack";
-import rulesLoaders from './webpack-config-scripts/loaders/rulesLoaders'
+import rulesLoaders from './webpack-config-scripts/rulesLoaders'
 import plugins from "./webpack-config-scripts/plugins";
+import optimization from "./webpack-config-scripts/optimization";
 import devServer from "./webpack-config-scripts/devServer";
 import {fileName, isDev, __dirname} from "./webpack-config-scripts/modes";
 
@@ -15,7 +16,7 @@ const config: webpack.Configuration = {
 	},
 	output: {
 		filename: `static/js/${fileName('js')}`,
-		path: path.resolve(__dirname, 'dist/'),
+		path: path.resolve(__dirname, '../dist/'),
 		assetModuleFilename: 'static/media/[name].[hash][ext]',
 		clean: true,
 	},
@@ -27,7 +28,7 @@ const config: webpack.Configuration = {
 		},
 	},
 	devServer: devServer,
-	// optimization: optimization(),
+	optimization: optimization(),
     plugins: plugins(),
 	module: {
 		rules: rulesLoaders()
